feat: error handling
